python-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

sentence-transformers==2.2.2
chromadb==0.4.18
langchain==0.0.350
langchain-community>=0.0.2

langdetect==1.0.9
pydantic==2.5.0
requests==2.31.0
